requirements.txt
amqp==1.4.9
anyjson==0.3.3
beautifulsoup4==4.8.2
billiard==3.3.0.23
celery==3.1.26.post2
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.1
confusable-homoglyphs==3.2.0
Django==2.2.9
django-admin-interface==0.12.0
django-apptemplates==1.4
django-celery==3.3.1
django-ckeditor==5.6.1
django-colorfield==0.1.15
django-crispy-forms==1.7.2
django-db-signals==0.1.1
django-filter==2.0.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-google-analytics-app==4.4.0
django-jet==1.0.8
django-js-asset==1.1.0
django-registration==3.0
django-wysiwyg==0.8.0
feedparser==5.2.1
google-api-python-client==1.4.1
httplib2==0.15.0
idna==2.8
kombu==3.0.37
mccabe==0.6.1
oauth2client==4.1.3
Pillow==5.3.0
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
pyasn1==0.4.8
pyasn1-modules==0.2.7
pytz==2018.7
requests==2.22.0
rsa==4.0
six==1.13.0
soupsieve==1.9.5
sqlparse==0.3.0
structlog==19.2.0
uritemplate==3.0.1
urllib3==1.25.7
